+++ app/Http/Controllers/UserWordController.php
@@ -16,8 +16,13 @@ class UserWordController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-     public function index() {
-     	return response()-> json(['message' => 'Request executed successfully', 'users_words'=>User_Word::orderBy('user_id')->get()],200);
+     public function index(User $user) 
+     {
+     	if (Auth::user()->permissions()->where('permission_id','1')->exists() || Auth::user()->children()->where('student_id', $user->id)->exists() || Auth::user() == $user) {
+     		$words = $user->words;
+     		return response()-> json(['message' => 'Request executed successfully', 'words'=>$words],200);-
+		}
+		return response()->json(['message'=>'You do not have the permissions to view this users information.', 'code'=>403],403);
      }
 
      /**
@@ -29,14 +34,17 @@ class UserWordController extends Controller
      */
     public function store(User $user, Request $request)
     {
-    	foreach ($request->all() as $word_id){
-            if ($word=Word::find($word_id)){
-                $word->users()->sync($user->id, false);
-            } else {
-                response()->json(['message'=>'Error in word chosen'], 401);
-            } 
-        }
-        return response()->json(['message' => 'Word(s) correctly added to user', 'words added'=>$user->words,'code'=>201]);
+    	if (Auth::user()->permissions()->where('permission_id','2')->exists() || Auth::user()->children()->where('student_id', $user->id)->exists() || Auth::user() == $user) {
+	    	foreach ($request->all() as $word_id){
+	            if ($word=Word::find($word_id)){
+	                $word->users()->sync($user->id, false);
+	            } else {
+	                response()->json(['message'=>'Error in word chosen'], 401);
+	            } 
+	        }
+	        return response()->json(['message' => 'Word(s) correctly added to user', 'words added'=>$user->words,'code'=>201]);
+	    }
+		return response()->json(['message'=>'You do not have the permissions to edit this users information.', 'code'=>403],403);
     }
 
      /**
@@ -48,12 +56,16 @@ class UserWordController extends Controller
      */
     public function show(User $user, Word $word)
     {
-    	$words = $user->words->find($word->id);
-        if (!$words) {
-            return response()->json(['message' => 'This user is currently not working on this word.', 'code'=>404], 404);
-        }
-        $user_word = $user->words()->whereWordId($word->id)->get();
-		return response()->json(['message' =>'Successful retrieval of record.', 'user_word'=>$user_word, 'code'=>201], 201);
+    	if (Auth::user()->permissions()->where('permission_id','1')->exists() || Auth::user()->children()->where('student_id', $user->id)->exists() || Auth::user() == $user) {
+    		$words = $user->words->find($word->id);
+	        if (!$words) {
+	            return response()->json(['message' => 'This user is currently not working on this word.', 'code'=>404], 404);
+	        }
+	        $user_word = $user->words()->whereWordId($word->id)->get();
+			return response()->json(['message' =>'Successful retrieval of record.', 'user_word'=>$user_word, 'code'=>201], 201);
+		}
+		return response()->json(['message'=>'You do not have the permissions to view this users information.', 'code'=>403],403);
+    	
     }
 
     /**
@@ -65,60 +77,72 @@ class UserWordController extends Controller
      */
     public function update(Request $request, User $user, Word $word)
     {
-    	$words = $user->words->find($word->id);
-        if (!$words) {
-            return response()->json(['message' => 'This user is currently not working on this word.', 'code'=>404], 404);
-        }
+    	if (Auth::user()->permissions()->where('permission_id','2')->exists() || Auth::user()->children()->where('student_id', $user->id)->exists() || Auth::user() == $user) {
+    		$words = $user->words->find($word->id);
+	        if (!$words) {
+	            return response()->json(['message' => 'This user is currently not working on this word.', 'code'=>404], 404);
+	        }
 
-        else if ($request->exists('user_id') || $request->exists('word_id')) {
-        	return response()->json(['message' => 'You cannot change these values.', 'code'=>401], 401);
-        }
+	        else if ($request->exists('user_id') || $request->exists('word_id')) {
+	        	return response()->json(['message' => 'You cannot change these values.', 'code'=>401], 401);
+	        }
 
-        $user_word = User_Word::where('user_id', $user->id)->where('word_id', $word->id)->first();
+	        $user_word = User_Word::where('user_id', $user->id)->where('word_id', $word->id)->first();
 
-        $user_word->fill($request->all())->save();
+	        $user_word->fill($request->all())->save();
 
-        return response()->json(['message'=>'Record updated','user_word' => $user_word, 201], 201);
+	        return response()->json(['message'=>'Record updated','user_word' => $user_word, 201], 201);
+	    }
+		return response()->json(['message'=>'You do not have the permissions to edit this users information.', 'code'=>403],403);
     }
 
     public function addView(User $user, Word $word)
     {
-    	$words = $user->words->find($word->id);
-        if (!$words) {
-            return response()->json(['message' => 'This user is currently not working on this word.', 'code'=>404], 404);
-        }
-        $user_word = User_Word::where('user_id', $user->id)->where('word_id', $word->id)->first(); 
-        
-        $user_word->view_count++;
+    	if (Auth::user()->permissions()->where('permission_id','2')->exists() || Auth::user()->children()->where('student_id', $user->id)->exists() || Auth::user() == $user) {
+    		$words = $user->words->find($word->id);
+	        if (!$words) {
+	            return response()->json(['message' => 'This user is currently not working on this word.', 'code'=>404], 404);
+	        }
+	        $user_word = User_Word::where('user_id', $user->id)->where('word_id', $word->id)->first(); 
+	        
+	        $user_word->view_count++;
 
-        $user_word->save();
+	        $user_word->save();
 
-        return response()->json(['message'=>'Record updated','user_word' => $user_word, 201], 201);
+	        return response()->json(['message'=>'Record updated','user_word' => $user_word, 201], 201);
+	    }
+    	return response()->json(['message'=>'You do not have the permissions to edit this users information.', 'code'=>403],403);
     }
 
 	public function toggleActive(User $user, Word $word)
     {
-    	$words = $user->words->find($word->id);
-        if (!$words) {
-            return response()->json(['message' => 'This user is currently not working on this word.', 'code'=>404], 404);
-        }
-        $user_word = User_Word::where('user_id', $user->id)->where('word_id', $word->id)->first(); 
-        
-        $user_word->active = !($user_word->active);
+    	if (Auth::user()->permissions()->where('permission_id','2')->exists() || Auth::user()->children()->where('student_id', $user->id)->exists() || Auth::user() == $user) {
+    		$words = $user->words->find($word->id);
+	        if (!$words) {
+	            return response()->json(['message' => 'This user is currently not working on this word.', 'code'=>404], 404);
+	        }
+	        $user_word = User_Word::where('user_id', $user->id)->where('word_id', $word->id)->first(); 
+	        
+	        $user_word->active = !($user_word->active);
 
-        $user_word->save();
+	        $user_word->save();
 
-        return response()->json(['message'=>'Record updated','user_word' => $user_word, 201], 201);
+	        return response()->json(['message'=>'Record updated','user_word' => $user_word, 201], 201);
+	    }
+    	return response()->json(['message'=>'You do not have the permissions to edit this users information.', 'code'=>403],403);
     }
 
     public function newWord(User $user)
     {
-    	$word_ids = User_Word::select('word_id')->where('user_id',$user->id)->get();
-    	if ($new_word = Word::where('lexile_level', $user->lexile_level)->whereNotIn('id', $word_ids)->inRandomOrder()->first()){
-    		$new_word->users()->sync($user->id, false);
-    		return response()->json(['message' => 'Word(s) correctly added to user', 'current words'=>$user->words,'code'=>201]);
-    	}
-    	return response()->json(['message'=>'This user has no additional words of this lexile level.'], 401);
+    	if (Auth::user()->permissions()->where('permission_id','2')->exists() || Auth::user()->children()->where('student_id', $user->id)->exists() || Auth::user() == $user) {
+    		$word_ids = User_Word::select('word_id')->where('user_id',$user->id)->get();
+	    	if ($new_word = Word::where('lexile_level', $user->lexile_level)->whereNotIn('id', $word_ids)->inRandomOrder()->first()){
+	    		$new_word->users()->sync($user->id, false);
+	    		return response()->json(['message' => 'Word(s) correctly added to user', 'current words'=>$user->words,'code'=>201]);
+	    	}
+	    	return response()->json(['message'=>'This user has no additional words of this lexile level.'], 401);
+	    }
+    	return response()->json(['message'=>'You do not have the permissions to edit this users information.', 'code'=>403],403);
     }
 
     /**
@@ -129,11 +153,14 @@ class UserWordController extends Controller
      */
     public function destroy(User $user, Word $word)
     {
-    	$words = $user->words->find($word->id);
-        if (!$words) {
-            return response()->json(['message' => 'This user is currently not working on this word.', 'code'=>404], 404);
-        }
-        $user->words()->detach($word->id);
-        return response()->json(['message'=>'The user is no longer working on this word.', 'words'=>$user->words()->get(), 'code'=>201], 201);
+    	if (Auth::user()->permissions()->where('permission_id','2')->exists() || Auth::user()->children()->where('student_id', $user->id)->exists() || Auth::user() == $user) {
+    		$words = $user->words->find($word->id);
+	        if (!$words) {
+	            return response()->json(['message' => 'This user is currently not working on this word.', 'code'=>404], 404);
+	        }
+	        $user->words()->detach($word->id);
+	        return response()->json(['message'=>'The user is no longer working on this word.', 'words'=>$user->words()->get(), 'code'=>201], 201);
+		}
+    	return response()->json(['message'=>'You do not have the permissions to edit this users information.', 'code'=>403],403);
     }
 }
